agents/pyproject.toml
[project]
name = "project-agents"
version = "0.1.0"
requires-python = ">=3.12"
license = { text = "EUPL-1.2" }
dependencies = [
  "haystack-ai>=2.24,<2.25",
  "fastapi",
  "uvicorn[standard]",
  "python-dotenv",
  "httpx",
  "jinja2",
  "opentelemetry-sdk",
  "opentelemetry-exporter-otlp-proto-http",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-httpx",
]

[dependency-groups]
dev = [
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "pytest",
  "anyio",
  "trio",
  "httpx",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]
"app.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["^.venv/", "^.pytest_cache/"]
